The Great Reformation 3.
The project is now a leaps ahead more stable and uses a more rigid albeit not strictly rigid typing system via Pylance. Now the project should be more responsive to to request for the components of the engine.
Note: in near future grid.py will be made into an array type class which will have the desired features which I have implemented for grids.